LayoutTests/imported/w3c/web-platform-tests/css/selectors/invalidation/has-sibling-insertion-removal-expected.txt


PASS subject1: initial color should be rgb(128, 128, 128)
FAIL subject1: color after #sibling1_1 inserted should be rgb(255, 0, 0) assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS subject2: initial color should be rgb(128, 128, 128)
FAIL subject2: color after #sibling2_1 removed should be rgb(0, 128, 0) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS subject3: initial color should be rgb(128, 128, 128)
FAIL subject3: color after #sibling3_1 inserted should be rgb(0, 0, 255) assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS subject4: initial color should be rgb(128, 128, 128)
FAIL subject4: color after #sibling4_1 removed should be rgb(255, 255, 0) assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS subject5: initial color should be rgb(255, 0, 0)
FAIL subject5: color after #sibling5_1 removed should be rgb(128, 128, 128) assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
PASS subject6: initial color should be rgb(0, 128, 0)
FAIL subject6: color after #sibling6_1 inserted should be rgb(128, 128, 128) assert_equals: expected "rgb(128, 128, 128)" but got "rgb(0, 128, 0)"
PASS subject7: initial color should be rgb(0, 0, 255)
FAIL subject7: color after #sibling7_1 removed should be rgb(128, 128, 128) assert_equals: expected "rgb(128, 128, 128)" but got "rgb(0, 0, 255)"
PASS subject8: initial color should be rgb(255, 255, 0)
FAIL subject8: color after #sibling8_1 inserted should be rgb(128, 128, 128) assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 255, 0)"
PASS subject9: initial color should be rgb(128, 128, 128)
FAIL subject9: color after #sibling9_1 inserted should be rgb(255, 0, 0) assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS subject10: initial color should be rgb(128, 128, 128)
FAIL subject10: color after #sibling10_1 removed should be rgb(0, 128, 0) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS subject11: initial color should be rgb(128, 128, 128)
FAIL subject11: color after #sibling11_1 inserted should be rgb(0, 0, 255) assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS subject12: initial color should be rgb(128, 128, 128)
FAIL subject12: color after #sibling12_1 removed should be rgb(255, 255, 0) assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS subject13: initial color should be rgb(128, 128, 128)
FAIL subject13: color after #sibling12_1 removed should be rgb(0, 128, 0) assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS subject14: initial color should be rgb(0, 128, 0)
FAIL subject14: color after #sibling14_2 removed should be rgb(128, 128, 128) assert_equals: expected "rgb(128, 128, 128)" but got "rgb(0, 128, 0)"
